Corrections to call graph
Functions on the call graphs were indexed by names. When I recently started to analyze something it turned out that if I change the name of the function ("already know what this one does so let's name it this and this") the graph was still using the same one and was wrong. Fixed now, indexed by start address of function (watch out for rebasing though!)